Adding scores dynamically
